refactor: reorganize code structure by splitting lib.rs into specialized modules
- Create src/shm.rs for shared memory and red-black tree operations
  - Move vts_init_shm_zone and vts_rbtree_insert_value functions
  - Move VtsSharedContext struct definition
- Move utility functions to src/prometheus.rs
  - Move generate_vts_status_content, get_hostname, and get_current_time
- Update lib.rs imports and remove duplicated code
- Fix all clippy warnings and maintain proper formatting
- Verify nginx integration continues to work correctly

This refactoring improves code organization and maintainability while
preserving all existing functionality as specified in ISSUE7.md.
